OpenUSD Autoupgrade: Let it rip & begin damage control.

+++ Sources/Gf/include/Gf/size3.h
@@ -1,25 +1,8 @@
 //
 // Copyright 2016 Pixar
 //
-// Licensed under the Apache License, Version 2.0 (the "Apache License")
-// with the following modification; you may not use this file except in
-// compliance with the Apache License and the following modification to it:
-// Section 6. Trademarks. is deleted and replaced with:
-//
-// 6. Trademarks. This License does not grant permission to use the trade
-//    names, trademarks, service marks, or product names of the Licensor
-//    and its affiliates, except as required to comply with Section 4(c) of
-//    the License and to reproduce the content of the NOTICE file.
-//
-// You may obtain a copy of the Apache License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the Apache License with the above modification is
-// distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied. See the Apache License for the specific
-// language governing permissions and limitations under the Apache License.
+// Licensed under the terms set forth in the LICENSE.txt file available at
+// https://openusd.org/license.
 //
 #ifndef PXR_BASE_GF_SIZE3_H
 #define PXR_BASE_GF_SIZE3_H
@@ -27,11 +10,10 @@
 /// \file gf/size3.h
 /// \ingroup group_gf_LinearAlgebra
 
-#include <pxr/pxrns.h>
-
-#include "Arch/pxrinttypes.h"
+#include "Arch/inttypes.h"
 #include "Gf/api.h"
 #include "Gf/vec3i.h"
+#include "pxr/pxrns.h"
 
 #include <iosfwd>
 
@@ -210,8 +192,7 @@ class GfSize3 {
   /// Conversion to GfVec3i
   operator GfVec3i() const
   {
-    return GfVec3i(
-        static_cast<int>(_vec[0]), static_cast<int>(_vec[1]), static_cast<int>(_vec[2]));
+    return GfVec3i(_vec[0], _vec[1], _vec[2]);
   }
 
  private:
